refactor: queryPostBySlug
