--- packages/core/src/domains/posts/queryPostBySlug/index.ts
@@ -1,15 +1,0 @@
-import { client } from '$config/sanity';
-
-import { postQuery } from './query';
-import type { BlogPostBySlug } from './types';
-import { blogPostBySlugSchema } from './types';
-
-export async function queryPostBySlug(slug: string): Promise<BlogPostBySlug> {
-  const result = await client.fetch<BlogPostBySlug>(postQuery, {
-    slug,
-  });
-
-  return blogPostBySlugSchema.parse(result);
-}
-
-export type { BlogPostBySlug } from './types';
